Typen korjaus

+++ KamkByte3/KamkByte3/Code/ObjectRelatives/Mold/ObjectMold.cpp
@@ -25,7 +25,7 @@ float ObjectMold::getScale()
 	return ObjectScale;
 }
 
-void ObjectMold::setScale(int scale)
+void ObjectMold::setScale(float scale)
 {
 	ObjectScale = scale;
 }
