+++ KamkByte3/KamkByte3/Code/ObjectRelatives/Mold/ObjectMold.h
@@ -13,7 +13,7 @@ public:
 	void setDirectory(std::string path);
 
 	float getScale();
-	void setScale(int scale);
+	void setScale(float scale);
 	
 	float getLocationPoint(int i);	
 	void setLocationPoints(float ax, float ay, float bx, float by, float cx, float cy, float dx, float dy, float ex, float ey, float fx, float fy, float z, float alpha);
